Hosting/Joining Changes. Custom Games
-Changed appID to an Int64 to acmomidate for possible large values from
hashing custom games exe.
-Fixed Disconnet for both the Host and the Joined
-Number of friends in lobby only shows if a lobby exists.
-If the host DC's the Joined will disconnect as well after a few
minutes.
-Added protection from null messages causing host thread to die horribly

+++ Sapp/Sapp/Game.cs
@@ -14,7 +14,7 @@ namespace Sapp
         // maybe consolidate these attributes into a GameProperties object?
         
 //        private List<string> genre;
-        private int appID;
+        private Int64 appID;
         
         private bool isInstalled;
         private int lastTimePlayed;
@@ -77,7 +77,7 @@ namespace Sapp
  //           PopulateTags(tags);
         }
 
-        public int GetAppID()
+        public Int64 GetAppID()
         {
             return appID;
         }
